Add ExactDate expression
When working with calendars it is a common task to omit specific dates.
This commit adds an `ExactDate` expression that can model these
omissions very clearly.

Let's say you have a basketball practice that meets every Monday, except
on July 4 2016 because that is a holiday. With this new Expression you
could model this as:

```ruby
schedule_with_cancellations = Repeatable::Schedule.new({
  difference: {
    included: { weekday: { weekday: 1 } },
    excluded: { exact_date: { date: "2016-07-04" } },
  },
})
```