Use negative numbers for :last for DayInMonth
The use of a special :last symbol (or string) for defining an expression
for the last day of a month worked well.

Now that we have added the ability to do a similar thing for
WeekdayInMonth and ended up with a more Ruby-ish and more flexible
negative number system for describing the number of weeks (or days) from
the end of a month, it follows that we'd updated DayInMonth to match.

This gives us increased flexibility to define (in the somewhat rare case
that you'd need it) something like "3rd to last day of the month." More
importantly, it brings consistency to the interfaces of these classes
and removes a "magic" symbol/string.