Install sorbet and run setup
I'd like to see what it takes to get Sorbet running in this gem, partly
as a small experiment to try Sorbet in general, partly to provide type
signatures for other projects that may be using this gem and Sorbet, and
partly to see what sorts of benefits it will produce for the development
of the gem itself.

I don't anticipate any huge development gains or bug catching, but this
feels like a good small project to try Sorbet on. And... you never know
:)