Upgrade all Date expressions to strict typing
Getting into the nitty gritty of using Sorbet now. These changes require
the use of method signature and variable type annotations. This raised a
few questions about consistency in when we use conversion functions for
params and how to describe the types allowed by a method. But I decided
to make this as small a change as possible, simply noting what is
currently used in the methods themselves.

The big gotcha here is that Sorbet only allows declaration (i.e.
setting) of ivars within the `initialize` method. This meant that in the
`Biweekly` class, I had to move the work of `find_first_occurrence` into
`initialize`. I don't love this because doing that much work in
`initialize` feels like an anti-pattern. But only in the most extreme
cases is that method going to involve a significant amount of time, and
the point of these classes is the `include?` method, which performs the
same work anyway. It's mostly moving the order of operations on the
whole.