Use abstract & implementation sorbet features
Especially in a setup like this where there are multiple levels of
inheritance, I found this feature really nice! It fixes one of the main
difficulties with Ruby's duck-typing features: there's not a great way
to enforce that duck-typing. This lets us describe the abstract classes
and which methods in them are abstract and enforces that implementations
of those methods by subclasses conform to the same (or contravariant)
type signatures.