Flatten nested Intersection expressions
As is true with `Unions`, nested `Intersection` elements are equivalent
to their flattened representation.  With this implementation, there's
not a good way to prove that equivalence in tests.  But the set theory
bears this out.

This mirrors the implementation for `Union`, and I did consider bumping
this implementation up to their shared parent, `Set`. But that feels a
little misleading, especially if any other `Set` classes come around for
which doesn't share this particular characteristic of `Union`s and
`Intersection`s.  I don't know what that set would be, but I think it's
worth holding off for now. If we need to copy this a third time, then we
can find the right abstraction (:hat_tip: Sandi Metz).