Turn on strict type checking for Date expressions
Strict type checking requires that we have a type signature for every
method in a file and all ivar declarations.

The only real noteworthy thing here is that Biweekly#first_occurrence
moved from a memoized method to having its value set in initialize. This
doesn't seem ideal from a pure Ruby performance perspective (i.e. leave
more expensive calculations until we actually need them), but I think
it's fine here since the calculation isn't that expensive and this value
is needed for most method calls on this class. All the same, I'd love to
revisit this as a possibility as a memoized method in the future.