Turn on strict typing for Set expressions
As with the Date expressions, this was pretty straightforward with one
exception. Because of Sorbet's poor support for splat operators, the
setup we have where Intersection and Union do some optimization in their
initialize methods, the types are a little funky for Set#initialize. The
way I have it now is the easiest to parse but may not reflect reality in
all situations.  I'm going to circle back with some tests to try and
exercise these boundaries. The big bummer is that defining things in the
easiest way for Sorbet (take a single array argument) would be a
breaking change.  But maybe we can figure something out.