Fix biweekly first_occurrence value in to_h output
It was never intended that the `first_occurrence` would be part of the
hash representation of a biweekly expression. For the most part, it
wasn't but this failing test showed that we not only have a bug, but
we've had a bug for a while in that this `to_h` output would previously
(i.e. before setting `@first_occurrence` in `initialize`) have been
different for this class before and after calling `include?`. Before
that call, the ivar was not set, but after the call, it is set.

I chose to go the route of establishing a convention where prepending
any ivar with an underscore would effectively make it "private" as far
as the `to_h` method is concerned. This works well and seems pretty
clear.