Use sorbet abstract helpers
Sorbet's abstract and interface class helpers are extremely useful for
more clearly defining these types of classes in Ruby where we normally
have to rely on things like `NotImplementedError` and shared examples to
make sure that each implementation of a different type or interface
actually defines that interface. It's worth reading the documentation
about these helpers.

In this change we set up a few abstract classes as those that in our
hierarchy of Expressions shouldn't be instantiated on their own but have
some shared methods and outline other methods in the shared interface of
what defines an "expression".