Upgrade Schedule to strict typing
This should mostly be helpful for documentation purposes. Defining these
method signatures and such highlighted an edge case that likely wouldn't
be reached but was technically possible in #next_occurrence. In
addressing the issue, I think it made the looping functionality and when
we break out of it a little more clear than it was before, which feels
like a win-win.