Stop including module in top level namespace
This worked ok but isn't the most considerate way to handle this as a
library. This is a breaking change if an app or other library using
this gem relied on access to this Date() method. They can still access
it through the more verbose Repeatable::Conversions::Date(), and they
can always include Repeatable::Conversions in their own top level
namespace if they want. But this is the better path forward.