Add standard for formatting
Style has been fairly consistent with this project, but with the
existence of a few tools to automatically format our code, the hope is
that I and other contributors don't have to think about style. We can
write the code and let standard do the formatting.

I chose standard mostly because it removes any discussion about what
style rules to choose. I really don't care what the rules are, and using
standard, which isn't configurable, removes the need for those decisions
and discussions. The rubocop config it comes with is sensible enough,
and the consistency will be a win.

Right now, this is setup in the way that the creator of standard
suggests for gems (https://git.io/JYTVn), which includes adding its
check to the default rake task. This means, for us and in most general
cases, that the check will be run in CI. That'll at least let us see
code that needs formatting so we can address it. Ideally, I'd like
something to run this check with --fix on all PRs automatically, but
I'll leave that for another day.