Use default CI script
For Travis CI, this is `rake`, which we already have setup to run RSpec.
So there's no functional change just yet, but that lets us add other
important bits of tooling to the default rake task and not have to worry
about what command to run. When in doubt, just run `bundle exec rake`!