Add smoke test for runtime checks
It's possible to disable sorbet runtime checks in a variety of
scenarios. Both as a way of demonstrating the desired behavior and
protecting from regressions here, we can write a test that exercises
sending the wrong type to a typed method.

Leaving runtime checks on :always is encouraged by the sorbet team, and
any library or application that includes this gem has the option of
changing the default check level, so we're not necessarily forcing
anything on others.

Additionally, the main public methods, especially for Schedule, have
been written in a way to be very permissive, doing a good deal of
conversion work internally. This feels like a nice sweet spot where we
can ensure that types are as they should be where the code isn't written
more defensively, but we don't enforce total rigidity.

I even toyed with the idea of changing `Schedule#initialize`'s
signature to only allow an Expression::Base or a Hash. But because we
already had some type checking in place, and the existing error message
seems markedly more friendly than the TypeError from sorbet, I think
it's worth leaving the existing implementation for such a key entry
point into the gem.