Ensure all Set classes can accept an array
Previous to the addition of sorbet, all Set classes could receive
elements for initialize as either separate arguments or an array of
arguments. The Set class retained this information in its types, but
Union and Intersection became more strict.

This change adds tests to demonstrate the supported use of arrays and
fixes the type checking errors that result.