feat(Daily): add new expression
We're using Repeatable to build a task manager, and in that task
manager, we want to support daily repeating tasks.

Our current implementation looks like this:

```ruby
daily = Repeatable::Schedule.new(
  union: [
    { weekday: { weekday: 0 } },
    { weekday: { weekday: 1 } },
    { weekday: { weekday: 2 } },
    { weekday: { weekday: 3 } },
    { weekday: { weekday: 4 } },
    { weekday: { weekday: 5 } },
    { weekday: { weekday: 6 } },
  ]
)
```

While the `Daily` expression is extremely simple, it is utile, and
allows for simpler storage and expression of this pattern.