Move Set specs to shared examples
In more recent versions of Sorbet, instantiating a class that's marked
as abstract will raise an exception, so all of our specs that test the
Set class directly are now failing. We can get the same functionality by
moving those to shared examples and including them in the Intersection
and Union specs. This required very little to change in the specs
themselves.