Fix type error for Date conversion method
Now that we have more type information for core and stdlib Ruby classes,
Sorbet is complaining that we're potentially sending things that aren't
Strings to Date.parse (whose first argument is said to be a String).

From my POV, this is a problem in the Sorbet type system because marking
the first argument for Date#parse as a String doesn't cover some
built-in functionality for that method.

If you use a non-String as the first argument, you'll get an error
complaining about that type not having an "implicit conversion to
String." This is a big feature of a number of Ruby parsing & conversion
methods. Classes can signal that they can behave and be treated just
like a String by responding to #to_str (this is different from the #to_s
method, which is _explicit_ conversion). But that functionality isn't
captured in the type system or by including a certain class.

This is a feature that I think is worth continuing to support for a
conversion method.

As this branch is the fallback branch and we intend to catch errors
where non-String and non-String-like objects are passed to Date#parse,
we'll use the T.let(object, T.untyped) escape hatch to keep our intended
functionality while appeasing Sorbet.

I hope that one day Sorbet can support interfaces that revolve around
responding to certain methods rather than class hierarchy alone. Until
then, this will do for us.