Work around *args limitations in Sorbet
First, for more recent versions of Sorbet, I had the type wrong for
these splatted arguments (i.e. *elements). The type should be the type
for each of the individual args, not the group of them.

Once that was sorted out, we hit a new problem. Calling super in
Intersection and Union initialize methods seemed to confuse the type
system into thinking that it was passing along the type wrapped in an
extra array.

Normally, we could splat the elements before calling super, but that led
to error 7019[^1], which pretty plainly says that splats aren't very
well supported in Sorbet right now. The suggestion is to have the
methods take an Array instead of splatted arguments.

Since the tests explicitly cover these methods accepting both a single
array and separate individual arguments, I'd rather not break that
functionality for this type issue. Perhaps we'll go that route in the
future, but for now we can fall back on the fact that Set is an abstract
class. We can change it to take only a single Array argument since
nothing should be calling it directly (and if it were, they'd be getting
a type error). But we'll leave the child classes as accepting splatted
args.

Also, because Sorbet doesn't quite understand how Array() works (i.e. if
you pass it an Array, it won't wrap that but return it as-is), we have
to do this Array(object).flatten dance to make sure that Sorbet is clear
on what we're dealing with.

Again, I don't love this, but it works for now. The type system is happy
and all the tests are passing.

[^1]: https://sorbet.org/docs/error-reference#7019